requirements.txt
altgraph==0.17.3
apscheduler==3.6.3
cartopy==0.21.1
cryptography==39.0.1
cycler==0.11.0
future==0.18.3
geos==3.8.0
kiwisolver==1.4.4
matplotlib-base==3.4.3
numpy==1.23.5
pycparser==2.21
pyparsing==3.0.9
pyproj==3.5.0
pyshp==2.3.1
python-dateutil==2.8.2
python-telegram-bot==13.4.1
pytz==2022.7
pywin32==305
pywin32-ctypes==0.2.0
shapely==1.8.4
six==1.16.0
tornado==6.2
tzlocal==2.1
ujson==5.4.0
wheel==0.38.4
